# docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Personal Style Network API",
    "version": "1.0.0",
    "description": "API cho ứng dụng Personal Style Network",
    "contact": {
      "name": "PSN Team",
      "email": "support@psn.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3052/v1/api",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/style-rules": {
      "get": {
        "summary": "Lấy danh sách quy tắc phong cách",
        "tags": [
          "StyleRules"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "ruleType",
            "schema": {
              "type": "string"
            },
            "description": "Loại quy tắc"
          },
          {
            "in": "query",
            "name": "isActive",
            "schema": {
              "type": "boolean"
            },
            "description": "Trạng thái kích hoạt"
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách quy tắc phong cách"
          },
          "401": {
            "description": "Không được ủy quyền"
          }
        }
      },
      "post": {
        "summary": "Tạo quy tắc phong cách mới",
        "tags": [
          "StyleRules"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "description",
                  "ruleType"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "ruleType": {
                    "type": "string",
                    "enum": [
                      "color_combination",
                      "pattern_matching",
                      "proportion",
                      "occasion",
                      "body_type",
                      "season",
                      "general"
                    ]
                  },
                  "conditions": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "attribute": {
                          "type": "string"
                        },
                        "operator": {
                          "type": "string",
                          "enum": [
                            "equals",
                            "not_equals",
                            "greater_than",
                            "less_than",
                            "contains",
                            "not_contains"
                          ]
                        },
                        "value": {
                          "type": "object"
                        }
                      }
                    }
                  },
                  "recommendations": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "enum": [
                            "replace",
                            "add",
                            "remove",
                            "change_color",
                            "general"
                          ]
                        },
                        "description": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "isPublic": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tạo quy tắc phong cách thành công"
          },
          "400": {
            "description": "Dữ liệu không hợp lệ"
          },
          "401": {
            "description": "Không được ủy quyền"
          }
        }
      }
    },
    "/style-rules/{id}": {
      "get": {
        "summary": "Lấy thông tin quy tắc phong cách",
        "tags": [
          "StyleRules"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID của quy tắc phong cách"
          }
        ],
        "responses": {
          "200": {
            "description": "Thông tin quy tắc phong cách"
          },
          "401": {
            "description": "Không được ủy quyền"
          },
          "404": {
            "description": "Không tìm thấy quy tắc phong cách"
          }
        }
      },
      "put": {
        "summary": "Cập nhật quy tắc phong cách",
        "tags": [
          "StyleRules"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID của quy tắc phong cách"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "conditions": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  },
                  "recommendations": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  },
                  "isActive": {
                    "type": "boolean"
                  },
                  "ruleType": {
                    "type": "string",
                    "enum": [
                      "color_combination",
                      "pattern_matching",
                      "proportion",
                      "occasion",
                      "body_type",
                      "season",
                      "general"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cập nhật quy tắc phong cách thành công"
          },
          "400": {
            "description": "Dữ liệu không hợp lệ"
          },
          "401": {
            "description": "Không được ủy quyền"
          },
          "404": {
            "description": "Không tìm thấy quy tắc phong cách"
          }
        }
      },
      "delete": {
        "summary": "Xóa quy tắc phong cách",
        "tags": [
          "StyleRules"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID của quy tắc phong cách"
          }
        ],
        "responses": {
          "200": {
            "description": "Xóa quy tắc phong cách thành công"
          },
          "401": {
            "description": "Không được ủy quyền"
          },
          "404": {
            "description": "Không tìm thấy quy tắc phong cách"
          }
        }
      }
    },
    "/style-rules/evaluate": {
      "post": {
        "summary": "Đánh giá trang phục theo quy tắc phong cách",
        "tags": [
          "StyleRules"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "outfitId"
                ],
                "properties": {
                  "outfitId": {
                    "type": "string"
                  },
                  "ruleTypes": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "color_combination",
                        "pattern_matching",
                        "proportion",
                        "occasion",
                        "body_type",
                        "season",
                        "general"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đánh giá trang phục thành công"
          },
          "400": {
            "description": "Dữ liệu không hợp lệ"
          },
          "401": {
            "description": "Không được ủy quyền"
          },
          "404": {
            "description": "Không tìm thấy trang phục"
          }
        }
      }
    },
...
  },
...
}